Fix fallout from namespacing.
It appears the namespacing in 7e7fe96209c133e290a7fd44310dc7fdffdc6d70
is the cause of this problem.

When we're in the `ManageIQ::NetworkDiscovery::DiscoverProbe` module namespace,
we cannot do `Object.const_get("IpmiProbe")` and have it find
`ManageIQ::NetworkDiscovery::IpmiProbe`.  Instead, we need to look in the
`ManageIQ::NetworkDiscovery` namespace for `IpmiProbe`.

This worked previously, because IpmiProbe was at the toplevel namespace.